Add monthly listening stats for users
Adds a new entry into the user info embed, displaying the length of time
a user has listened to music in the current month. To accomplish this,
two new fields have been added to the database.

Integration steps: To migrate an existing database, these two new
columns will need to be manually added in SQLite, and ideally
initialized for all existing users to 0.