Improve Docker environment
Changes the Docker environment to use uv rather than pip. The Docker
image also includes the bot source code, rather than requiring a checked
out version of the repository to be included in a volume.

Despite using uv, an autogenerated requirements.txt is still maintained
for users who perfer to develop with pip.